+++ web/packages/backend/src/app.module.ts
@@ -2,8 +2,10 @@ import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo'
 import { Module } from '@nestjs/common'
 import { GraphQLModule } from '@nestjs/graphql'
 import { JwtService } from '@nestjs/jwt'
-import { AuthController, AuthService } from './modules/auth'
-import { HealthController, HealthService } from './modules/health'
+import { AuthController } from './modules/auth/auth.controller'
+import { AuthService } from './modules/auth/auth.service'
+import { HealthController } from './modules/health/health.controller'
+import { HealthService } from './modules/health/health.service'
 import { TicketModule } from './modules/ticket/ticket.module'
 
 @Module({
