--- web/packages/backend/src/modules/auth/index.ts
@@ -1,2 +1,0 @@
-export * from './controller'
-export * from './service'
