+++ web/packages/backend/src/modules/health/health.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common'
-import { HealthService } from './service'
+import { HealthService } from './health.service'
 
 @Controller()
 export class HealthController {
